+++ docs/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.1] - 2023-07-20
+### Changed
+- Fix `FileNotFoundError` when caching for different languages.
+- Use `targets` in index file for searching.
+
 ## [0.8.0] - 2023-07-15
 ### Added
 - Support for language and platform in config file.